requirements.txt
alembic==1.8.1
altair==5.4.1
annotated-types==0.6.0
anyio==3.6.2
apache-airflow==2.4.2
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.1
apispec==3.3.2
appdirs==1.4.4
argcomplete==2.0.0
argparse-ext==1.4.2
arrow==1.3.0
astroid==2.9.3
attrs==22.2.0
Babel==2.10.3
blinker==1.7.0
build==0.10.0
cachelib==0.9.0
cachetools==5.2.0
cattrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
cloudevents==1.9.0
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
coverage==7.2.0
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==38.0.1
db-dtypes==1.3.0
Deprecated==1.2.13
deprecation==2.1.0
diff-cover==7.0.1
dill==0.3.6
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.0.0
Faker==18.6.1
fiscalyear==0.4.0
Flask==3.0.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatten-json==0.1.14
functions-framework==3.2.1
geneimpacts==0.3.7
geographiclib==2.0
geonamescache==1.5.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.17
google-api-core==2.24.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-appengine-logging==1.1.6
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.27.0
google-cloud-bigquery-datatransfer==3.11.1
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.4.1
google-cloud-data-qna==0.10.13
google-cloud-error-reporting==1.6.0
google-cloud-iam==2.12.0
google-cloud-logging==3.5.0
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-generativeai==0.4.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
graphlib==0.9.5
graphviz==0.20.1
greenlet==1.1.3.post0
grpc-google-iam-v1==0.12.6
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==22.0.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
inheritance==0.1.5
iniconfig==2.0.0
ipython-genutils==0.2.0
isort==5.12.0
itsdangerous==2.1.2
jinja-cli==1.2.2
Jinja2==3.1.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
mdit-py-plugins==0.3.1
mdurl==0.1.2
narwhals==1.13.1
netifaces==0.11.0
numpy==1.23.4
oauthlib==3.2.2
openpyxl==1.8.6
packaging==23.0
pandas==2.2.3
pandas-gbq==0.25.0
pathlib==1.0.1
pathspec==0.9.0
pendulum==2.1.2
pgeocode==0.4.0
pickleshare==0.7.5
pillow==10.2.0
pip-tools==6.12.2
platformdirs==3.0.0
pluggy==1.0.0
prison==0.2.1
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.3
ptyprocess==0.7.0
py==1.11.0
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydata-google-auth==1.7.0
pydeck==0.9.1
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.12.2
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.18.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-env==0.6.2
pytest-mock==3.10.0
python-daemon==2.3.2
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python-sql==1.5.2
pytz==2022.5
pytzdata==2020.1
PyYAML==6.0
referencing==0.35.1
regex==2021.10.23
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
rpds-py==0.18.1
rsa==4.9
setproctitle==1.3.2
simplegeneric==0.8.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.42
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlfluff==1.3.2
sqlparse==0.4.3
streamlit==1.39.0
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tblib==1.7.0
tenacity==8.1.0
termcolor==2.0.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.4.1
tqdm==4.64.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.2
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.24.3
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.13.3
WTForms==3.0.1
xmltodict==0.13.0
yapf==0.32.0
zipp==3.10.0
